requirements.txt
click==8.0.4
DateTime==4.4
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2022.1
SQLAlchemy==1.4.32
uuid==1.30
Werkzeug==2.0.3
zope.interface==5.4.0
